+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - PWA support to allow installing app to mobile
+- right click context menu for habits
+- Pomodoro clock for habits
 
 ### Fixed